# commands: Reduce memory allocations in ZAdd