When adding lots of data in Redis, the allocations in ZAdd can put quite some pressure on the GC... Using the benchmark in https://gist.github.com/abustany/8f2cf2ee3b968020b75e , I get the following results:

```
benchmark              old ns/op     new ns/op     delta
BenchmarkRedisZAdd     22371         21779         -2.65%

benchmark              old allocs     new allocs     delta
BenchmarkRedisZAdd     10             9              -10.00%

benchmark              old bytes     new bytes     delta
BenchmarkRedisZAdd     373           343           -8.04%
```

There are probably a couple of places in the library where we could use a sync.Pool, for example for the commands, to reduce GC pressure, but that'd be the subject of another pull request...