@vmihailenco 2015-05-11
Looks good, but please add benchmark from https://gist.github.com/abustany/8f2cf2ee3b968020b75e to the redis_test.go file.