@vmihailenco 2015-05-15
Thanks. I rebased your changes and created new PR https://github.com/go-redis/redis/pull/108. Closing this one.